+++ doc/release_notes.rst
@@ -44,7 +44,11 @@ Upcoming Release
   network has been moved from ``focus_weights:`` to ``clustering:
   focus_weights:``. Backwards compatibility to old config files is maintained.
 
+* Add VOM as marginal cost to PtX processes.
+
 * The ``mock_snakemake`` function can now be used with a Snakefile from a different directory using the new ``root_dir`` argument.
+
+* Add option to capture CO2 contained in biogas when upgrading (``sector: biogas_to_gas_cc``).
 
 * Merged option to extend geographical scope to Ukraine and Moldova. These
   countries are excluded by default and is currently constrained to power-sector